Move even more things out of main.rs
