+++ src-tauri/src/commands/mod.rs
@@ -1,4 +1,6 @@
 pub mod archive;
 pub mod cores;
+pub mod files;
 pub mod firmware;
 pub mod plugins;
+pub mod saves;
