+++ src-tauri/src/firmware.rs
@@ -4,9 +4,7 @@ use std::{path::PathBuf, time::Instant};
 use tauri::Emitter;
 use tokio::io::AsyncWriteExt;
 
-use crate::{
-    PocketSyncState, hashes::md5_for_file, progress::ProgressEvent, util::progress_download,
-};
+use crate::{hashes::md5_for_file, progress::ProgressEvent, util::progress_download};
 
 #[derive(Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
