+++ src-tauri/src/main.rs
@@ -3,37 +3,27 @@
     windows_subsystem = "windows"
 )]
 
-use async_walkdir::{DirEntry, WalkDir};
 use checks::{check_if_folder_looks_like_pocket, connection_task};
-use clean_fs::find_dotfiles;
 use extism::CancelHandle;
-use file_cache::{clear_file_caches, get_file_with_cache};
+use file_cache::clear_file_caches;
 use file_locks::FileLocks;
-use futures::stream::{self, StreamExt};
-use hashes::crc32_for_file;
 use install_zip::start_zip_task;
 use job_id::{Job, JobState};
-use log::{LevelFilter, debug, error, trace};
+use log::{LevelFilter, debug, trace};
 use root_files::RootFile;
 use save_sync_session::start_mister_save_sync_session;
-use saves_zip::{
-    SaveZipFile, build_save_zip, read_save_zip_list, read_saves_in_folder, read_saves_in_zip,
-    remove_leading_slash, restore_save_from_zip,
-};
+use saves_zip::SaveZipFile;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::path::PathBuf;
-use std::time::SystemTime;
 use std::vec;
-use tauri::{App, Emitter, Manager, RunEvent, Window};
+use tauri::{App, Emitter, Manager, RunEvent};
 use tauri_plugin_dialog::DialogExt;
 use tauri_plugin_log::{Target, TargetKind};
-use tokio::io::{AsyncReadExt, AsyncWriteExt};
 use tokio::sync::{Mutex, RwLock};
 
 use crate::app_error::AppError;
 use crate::hashes::{HashCache, HashCacheState};
-use crate::util::{find_common_path, get_mtime_timestamp};
 
 mod checks;
 mod clean_fs;
@@ -64,12 +54,10 @@ struct InnerState {
     file_locker: FileLocks,
     active_plugin_handles: Mutex<HashMap<String, CancelHandle>>,
     jobs: JobState,
-    hash_cache: RwLock<HashCache>,
 }
 
 struct PocketSyncState(InnerState);
 
-// Learn more about Tauri commands at https://tauri.app/v1/guides/features/command
 #[tauri::command(async)]
 async fn open_pocket(
     state: tauri::State<'_, PocketSyncState>,
@@ -111,381 +99,6 @@ async fn open_pocket_folder(
 }
 
 #[tauri::command(async)]
-async fn read_binary_file(
-    state: tauri::State<'_, PocketSyncState>,
-    path: &str,
-    app_handle: tauri::AppHandle,
-) -> Result<Vec<u8>, String> {
-    debug!("Command: read_binary_file - {path}");
-    let pocket_path = state.0.pocket_path.read().await;
-    let path = pocket_path.join(path);
-
-    let arc_lock = state.0.file_locker.find_lock_for(&path).await;
-    let _read_lock = arc_lock.read().await;
-
-    if let Ok(mut f) = if let Ok(cache_dir) = app_handle.path().app_cache_dir() {
-        get_file_with_cache(&path, &cache_dir).await
-    } else {
-        tokio::fs::File::open(&path).await
-    } {
-        let mut buffer = vec![];
-        f.read_to_end(&mut buffer)
-            .await
-            .expect(&format!("failed to read file: {:?}", path));
-
-        Ok(buffer)
-    } else {
-        Err(format!("No file found: {}", path.display()))
-    }
-}
-
-#[tauri::command(async)]
-async fn read_text_file(
-    state: tauri::State<'_, PocketSyncState>,
-    path: &str,
-    app_handle: tauri::AppHandle,
-) -> Result<String, ()> {
-    debug!("Command: read_text_file - {path}");
-    let pocket_path = state.0.pocket_path.read().await;
-    let path = pocket_path.join(path);
-
-    let arc_lock = state.0.file_locker.find_lock_for(&path).await;
-    let _read_lock = arc_lock.read().await;
-
-    let mut f = if let Ok(cache_dir) = app_handle.path().app_cache_dir() {
-        get_file_with_cache(&path, &cache_dir).await
-    } else {
-        tokio::fs::File::open(&path).await
-    }
-    .expect(&format!("no file found: {:?}", &path));
-
-    let mut file_contents = String::new();
-    f.read_to_string(&mut file_contents)
-        .await
-        .expect(&format!("failed to read file: {:?}", path));
-    Ok(file_contents)
-}
-
-#[tauri::command(async)]
-async fn file_exists(
-    state: tauri::State<'_, PocketSyncState>,
-    path: &str,
-) -> Result<bool, AppError> {
-    trace!("Command: file_exists - {path}");
-    let pocket_path = state.0.pocket_path.read().await;
-    let path = pocket_path.join(path);
-
-    let exists = tokio::fs::try_exists(&path).await?;
-    Ok(exists)
-}
-
-#[tauri::command(async)]
-async fn save_file(
-    path: &str,
-    buffer: Vec<u8>,
-    state: tauri::State<'_, PocketSyncState>,
-) -> Result<bool, ()> {
-    debug!("Command: save_file - {path}");
-    let file_path = PathBuf::from(path);
-    let folder_path = file_path.parent().unwrap();
-    let arc_lock = state.0.file_locker.find_lock_for(&file_path).await;
-    let _write_lock = arc_lock.write().await;
-    tokio::fs::create_dir_all(&folder_path).await.unwrap();
-    let mut file = tokio::fs::File::create(file_path).await.unwrap();
-    file.write_all(&buffer).await.unwrap();
-    file.flush().await.unwrap();
-    Ok(true)
-}
-
-#[tauri::command(async)]
-async fn list_files(
-    path: &str,
-    state: tauri::State<'_, PocketSyncState>,
-) -> Result<Vec<String>, ()> {
-    debug!("Command: list_files - {path}");
-    let pocket_path = state.0.pocket_path.read().await;
-    let dir_path = pocket_path.join(path);
-
-    let arc_lock = state.0.file_locker.find_lock_for(&dir_path).await;
-    trace!("list_files lock requested");
-    let _read_lock = arc_lock.read().await;
-    trace!("list_files lock granted");
-
-    if !tokio::fs::try_exists(&dir_path).await.unwrap() {
-        return Ok(vec![]);
-    }
-
-    let mut paths = tokio::fs::read_dir(dir_path).await.unwrap();
-    let mut results: Vec<_> = Vec::new();
-
-    while let Ok(Some(entry)) = paths.next_entry().await {
-        let file_type = entry.file_type().await.unwrap();
-        if file_type.is_file() {
-            let file_name = entry.file_name();
-            let file_name = file_name.to_str().unwrap();
-
-            if !file_name.starts_with(".") {
-                results.push(String::from(file_name))
-            }
-        }
-    }
-
-    Ok(results)
-}
-
-#[tauri::command(async)]
-async fn list_folders(
-    path: &str,
-    state: tauri::State<'_, PocketSyncState>,
-) -> Result<Vec<String>, ()> {
-    debug!("Command: list_folders - {path}");
-    let pocket_path = state.0.pocket_path.read().await;
-    let dir_path = pocket_path.join(path);
-
-    let arc_lock = state.0.file_locker.find_lock_for(&dir_path).await;
-    let _read_lock = arc_lock.read().await;
-
-    if !tokio::fs::try_exists(&dir_path).await.unwrap() {
-        return Ok(vec![]);
-    }
-
-    let mut paths = tokio::fs::read_dir(dir_path).await.unwrap();
-    let mut results: Vec<_> = Vec::new();
-
-    while let Ok(Some(entry)) = paths.next_entry().await {
-        let file_type = entry.file_type().await.unwrap();
-        if file_type.is_dir() {
-            let file_name = entry.file_name();
-            let file_name = file_name.to_str().unwrap();
-
-            if !file_name.starts_with(".") {
-                results.push(String::from(file_name))
-            }
-        }
-    }
-
-    Ok(results)
-}
-
-#[tauri::command(async)]
-async fn walkdir_list_files(
-    path: &str,
-    extensions: Vec<&str>,
-    off_pocket: Option<bool>,
-    state: tauri::State<'_, PocketSyncState>,
-) -> Result<Vec<String>, ()> {
-    debug!("Command: walkdir_list_files - {path}");
-    let pocket_path = state.0.pocket_path.read().await;
-    let dir_path = match off_pocket {
-        Some(true) => PathBuf::from(path),
-        None | Some(false) => pocket_path.join(remove_leading_slash(path)),
-    };
-
-    let arc_lock = state.0.file_locker.find_lock_for(&dir_path).await;
-    let _read_lock = arc_lock.read().await;
-
-    if !dir_path.exists() {
-        return Ok(vec![]);
-    }
-
-    fn is_hidden(entry: &DirEntry) -> bool {
-        entry
-            .file_name()
-            .to_str()
-            .map(|s| s.starts_with("."))
-            .unwrap_or(false)
-    }
-
-    let mut walker = WalkDir::new(&dir_path);
-    let dir_path_str = &dir_path.to_str().unwrap();
-    let mut file_paths = Vec::new();
-
-    while let Some(Ok(entry)) = walker.next().await {
-        match entry.file_type().await {
-            Ok(f) => {
-                if f.is_file() && !is_hidden(&entry) {
-                    let path = entry.path();
-                    let path_str = path.to_str().unwrap();
-                    let relative_path = path_str.replace(dir_path_str, "");
-                    if extensions.is_empty() || extensions.iter().any(|ext| path_str.ends_with(ext))
-                    {
-                        file_paths.push(relative_path);
-                    }
-                }
-            }
-            Err(_) => continue,
-        }
-    }
-
-    Ok(file_paths)
-}
-
-#[tauri::command(async)]
-async fn backup_saves(
-    save_paths: Vec<&str>,
-    zip_path: &str,
-    max_count: usize,
-    state: tauri::State<'_, PocketSyncState>,
-    hash_cache: tauri::State<'_, HashCacheState>,
-) -> Result<bool, ()> {
-    debug!("Command: backup_saves");
-    let pocket_path = state.0.pocket_path.read().await;
-    let hash_cache = hash_cache.inner();
-    build_save_zip(&pocket_path, save_paths, zip_path, max_count, hash_cache)
-        .await
-        .unwrap();
-
-    Ok(true)
-}
-
-#[tauri::command(async)]
-async fn list_backup_saves(
-    backup_path: &str,
-    hash_cache: tauri::State<'_, HashCacheState>,
-) -> Result<BackupSavesResponse, AppError> {
-    debug!("Command: list_backup_saves");
-    let path = PathBuf::from(backup_path);
-    if !path.exists() {
-        return Ok(BackupSavesResponse {
-            files: vec![],
-            exists: false,
-        });
-    }
-
-    let hash_cache = hash_cache.inner();
-    let files = read_save_zip_list(&path, hash_cache).await?;
-
-    Ok(BackupSavesResponse {
-        files,
-        exists: true,
-    })
-}
-
-#[tauri::command(async)]
-async fn list_saves_in_zip(zip_path: &str) -> Result<Vec<SaveZipFile>, AppError> {
-    debug!("Command: list_saves_in_zip");
-    let path = PathBuf::from(zip_path);
-    if !path.exists() {
-        return Ok(vec![]);
-    }
-
-    Ok(read_saves_in_zip(&path).await?)
-}
-
-#[tauri::command(async)]
-async fn list_saves_on_pocket(
-    state: tauri::State<'_, PocketSyncState>,
-    hash_cache: tauri::State<'_, HashCacheState>,
-) -> Result<Vec<SaveZipFile>, AppError> {
-    debug!("Command: list_saves_on_pocket");
-    let pocket_path = state.0.pocket_path.read().await;
-    let hash_cache = hash_cache.inner();
-    let saves_path = pocket_path.join("Saves");
-    Ok(read_saves_in_folder(&saves_path, Some(hash_cache)).await?)
-}
-
-#[tauri::command(async)]
-async fn restore_save(
-    zip_path: &str,
-    file_path: &str,
-    state: tauri::State<'_, PocketSyncState>,
-) -> Result<(), ()> {
-    debug!("Command: restore_save");
-    let pocket_path = state.0.pocket_path.read().await;
-    let path = PathBuf::from(zip_path);
-    restore_save_from_zip(&path, file_path, &pocket_path).await;
-
-    Ok(())
-}
-
-#[tauri::command(async)]
-async fn create_folder_if_missing(path: &str) -> Result<bool, ()> {
-    debug!("Command: create_folder_if_missing - {path}");
-    let folder_path = PathBuf::from(path);
-    if !folder_path.exists() {
-        tokio::fs::create_dir_all(path).await.unwrap();
-        return Ok(true);
-    }
-
-    Ok(false)
-}
-
-#[tauri::command(async)]
-async fn delete_files(
-    paths: Vec<&str>,
-    state: tauri::State<'_, PocketSyncState>,
-) -> Result<bool, ()> {
-    debug!("Command: delete_files");
-    let pocket_path = state.0.pocket_path.read().await;
-
-    let tasks: Vec<_> = paths
-        .into_iter()
-        .filter_map(|path| {
-            let file_path = pocket_path.join(path);
-            file_path
-                .exists()
-                .then(|| tokio::fs::remove_file(file_path))
-        })
-        .collect();
-
-    futures::future::join_all(tasks).await;
-    Ok(true)
-}
-
-#[tauri::command(async)]
-async fn copy_files(
-    copies: Vec<(&str, &str)>,
-    window: Window,
-    state: tauri::State<'_, PocketSyncState>,
-) -> Result<bool, ()> {
-    debug!("Command: copy_files");
-
-    let mut progress = progress::ProgressEmitter::new(Box::new(|event| {
-        window.emit("progress-event::copy_files", event).unwrap();
-    }));
-
-    progress.begin_work_units(copies.len());
-
-    let all_dests: Vec<PathBuf> = copies
-        .iter()
-        .map(|(_source, dest)| PathBuf::from(dest))
-        .collect();
-    let common_dir = find_common_path(&all_dests).unwrap();
-    let arc_lock = state.0.file_locker.find_lock_for(&common_dir).await;
-    let _write_lock = arc_lock.write().await;
-
-    for (origin, destination) in copies {
-        let origin = PathBuf::from(origin);
-        let destination = PathBuf::from(&destination);
-
-        if let Err(err) = match tokio::fs::create_dir_all(destination.parent().unwrap()).await {
-            Ok(_) => tokio::fs::copy(&origin, &destination).await,
-            Err(e) => Err(e),
-        } {
-            error!("{}", err);
-        } else {
-            progress.complete_work_units(1);
-            progress.set_message("file", Some(&destination.to_string_lossy()));
-        }
-    }
-
-    Ok(true)
-}
-
-#[tauri::command(async)]
-async fn find_cleanable_files(
-    path: &str,
-    state: tauri::State<'_, PocketSyncState>,
-) -> Result<Vec<String>, String> {
-    debug!("Command: find_cleanable_files");
-    let pocket_path = state.0.pocket_path.read().await;
-    let root_path = pocket_path.join(path);
-    let files = find_dotfiles(&root_path).await.unwrap();
-
-    Ok(files)
-}
-
-#[tauri::command(async)]
 async fn get_news_feed() -> Result<Vec<news_feed::FeedItem>, String> {
     debug!("Command: get_news_feed");
     let feed = news_feed::get_feed_json().await;
@@ -504,45 +117,6 @@ async fn begin_mister_sync_session(
         Ok(success) => return Ok(success),
         Err(err) => return Err(err.to_string()),
     }
-}
-
-#[tauri::command(async)]
-async fn get_file_metadata(
-    state: tauri::State<'_, PocketSyncState>,
-    hash_cache: tauri::State<'_, HashCacheState>,
-    file_path: &str,
-) -> Result<FileMetadata, AppError> {
-    trace!("Command: get_file_metadata");
-    let pocket_path = state.0.pocket_path.read().await;
-    let hash_cache = hash_cache.inner();
-    let full_path = pocket_path.join(file_path);
-
-    let crc32 = crc32_for_file(&full_path, Some(hash_cache)).await?;
-
-    let metadata = tokio::fs::metadata(full_path)
-        .await
-        .and_then(|m| m.modified())?;
-
-    let timestamp = metadata
-        .duration_since(SystemTime::UNIX_EPOCH)
-        .and_then(|d| Ok(d.as_secs()))?;
-
-    Ok(FileMetadata {
-        timestamp_secs: timestamp,
-        crc32,
-    })
-}
-
-#[tauri::command(async)]
-async fn get_file_metadata_mtime_only(
-    state: tauri::State<'_, PocketSyncState>,
-    file_path: &str,
-) -> Result<u64, AppError> {
-    trace!("Command: get_file_metadata_mtime_only");
-    let pocket_path = state.0.pocket_path.read().await;
-    let full_path = pocket_path.join(file_path);
-
-    Ok(get_mtime_timestamp(&full_path).await?)
 }
 
 #[tauri::command(async)]
@@ -566,28 +140,6 @@ async fn check_root_files(
     let pocket_path = state.0.pocket_path.read().await;
     let hash_cache = hash_cache.inner();
     Ok(root_files::check_root_files(&pocket_path, extensions, hash_cache).await?)
-}
-
-#[tauri::command(async)]
-async fn save_multiple_files(
-    state: tauri::State<'_, PocketSyncState>,
-    paths: Vec<&str>,
-    data: Vec<Vec<u8>>,
-) -> Result<(), AppError> {
-    debug!("Command: save_multiple_files");
-    let pocket_path = state.0.pocket_path.read().await;
-
-    let all_paths: Vec<PathBuf> = paths.iter().map(|p| pocket_path.join(p)).collect();
-    let common_dir = find_common_path(&all_paths).unwrap();
-
-    let arc_lock = state.0.file_locker.find_lock_for(&common_dir).await;
-    let _write_lock = arc_lock.write().await;
-
-    for (file_path, file_data) in all_paths.iter().zip(data) {
-        tokio::fs::write(file_path, file_data).await?;
-    }
-
-    Ok(())
 }
 
 #[tauri::command(async)]
@@ -640,45 +192,6 @@ async fn downconvert_single_pal_file(pal_file_path: String) -> Result<(), AppErr
     palettes::down_convert_pal_to_gbp(&pal_file_path).await?;
 
     Ok(())
-}
-
-#[tauri::command(async)]
-async fn find_mtime_for_files(full_file_paths: Vec<PathBuf>) -> Result<Vec<Option<u64>>, AppError> {
-    debug!("Command: find_mtime_for_files - {}", full_file_paths.len());
-    let paths_stream = stream::iter(full_file_paths);
-
-    let results = paths_stream
-        .map(|full_path| async move {
-            match get_mtime_timestamp(&full_path).await {
-                Ok(mtime) => Some(mtime),
-                Err(err) => {
-                    eprintln!("Error processing {:?}: {}", full_path, err);
-                    None // Return None for errors
-                }
-            }
-        })
-        .buffered(100)
-        .collect::<Vec<Option<u64>>>()
-        .await;
-
-    Ok(results)
-}
-
-#[tauri::command(async)]
-async fn get_folder_size(folder: PathBuf) -> Result<u128, AppError> {
-    debug!("Command: get_folder_size - {}", &folder.display());
-    let mut size = 0;
-    let mut walker = WalkDir::new(&folder);
-
-    while let Some(Ok(entry)) = walker.next().await {
-        if entry.path().is_file() {
-            if let Ok(meta) = entry.path().metadata() {
-                size += meta.len() as u128;
-            }
-        }
-    }
-
-    Ok(size)
 }
 
 #[tauri::command(async)]
@@ -766,44 +279,44 @@ fn main() {
         .invoke_handler(tauri::generate_handler![
             open_pocket,
             open_pocket_folder,
-            list_files,
-            list_folders,
-            walkdir_list_files,
-            read_binary_file,
-            read_text_file,
-            save_file,
+            commands::files::list_files,
+            commands::files::list_folders,
+            commands::files::walkdir_list_files,
+            commands::files::read_binary_file,
+            commands::files::read_text_file,
+            commands::files::save_file,
             commands::cores::uninstall_core,
             commands::archive::install_archive_files,
-            file_exists,
-            backup_saves,
-            list_backup_saves,
-            list_saves_in_zip,
-            list_saves_on_pocket,
-            restore_save,
-            create_folder_if_missing,
-            delete_files,
-            copy_files,
-            find_cleanable_files,
+            commands::files::file_exists,
+            commands::saves::backup_saves,
+            commands::saves::list_backup_saves,
+            commands::saves::list_saves_in_zip,
+            commands::saves::list_saves_on_pocket,
+            commands::saves::restore_save,
+            commands::files::create_folder_if_missing,
+            commands::files::delete_files,
+            commands::files::copy_files,
+            commands::files::find_cleanable_files,
             commands::cores::list_instance_packageable_cores,
             commands::cores::run_packager_for_core,
             get_news_feed,
             begin_mister_sync_session,
-            get_file_metadata,
-            get_file_metadata_mtime_only,
+            commands::files::get_file_metadata,
+            commands::files::get_file_metadata_mtime_only,
             commands::firmware::get_firmware_versions_list,
             commands::firmware::get_firmware_release_notes,
             commands::firmware::download_firmware,
             clear_file_cache,
             check_root_files,
             commands::archive::find_required_files,
-            save_multiple_files,
+            commands::files::save_multiple_files,
             get_active_jobs,
             stop_job,
             downconvert_all_pal_files,
             downconvert_single_pal_file,
-            find_mtime_for_files,
+            commands::files::find_mtime_for_files,
             move_game,
-            get_folder_size,
+            commands::files::get_folder_size,
             commands::plugins::run_plugin,
             commands::plugins::list_and_install_plugins,
             commands::plugins::uninstall_plugin,
